# api/openapi.yaml
openapi: 3.1.0
info:
  title: buffgo local control API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/watermarks:
    get:
      operationId: listWatermarks
      responses:
        "200":
          description: Short-pool minimums; usable is current short nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watermark"
    post:
      operationId: setWatermark
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatermarkSet"
      responses:
        "200":
          description: Watermark updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watermark"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/providers:
    get:
      operationId: listProviders
      responses:
        "200":
          description: Credential-safe vendor summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderCreate"
      responses:
        "201":
          description: Provider created; credential is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/providers/{id}/update:
    parameters:
      - $ref: "#/components/parameters/ProviderID"
    post:
      operationId: updateProvider
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderUpdate"
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/providers/{id}/credential:
    parameters:
      - $ref: "#/components/parameters/ProviderID"
    post:
      operationId: replaceProviderCredential
      security:
        - controlSession: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderCredentialReplace"
      responses:
        "200":
          description: Credential replaced; raw credential is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/providers/{id}/delete:
    parameters:
      - $ref: "#/components/parameters/ProviderID"
    post:
      operationId: deleteProvider
      security:
        - controlSession: []
          csrfToken: []
      responses:
        "200":
          description: Provider deleted
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    ProviderID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  securitySchemes:
    controlSession:
      type: apiKey
      in: cookie
      name: buffgo_control_session
    csrfToken:
      type: apiKey
      in: header
      name: X-Buffgo-CSRF
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
      properties:
        code:
          type: string
          enum:
            - invalid_host
            - method_not_allowed
            - origin_rejected
            - origin_required
            - json_required
            - csrf_rejected
            - internal_error
            - invalid_account_id
            - account_not_found
            - account_occupied
            - account_in_use
            - account_revision_conflict
            - session_required
            - session_too_large
            - invalid_json
            - body_not_allowed
            - resource_storage_unavailable
            - resource_integrity
            - resource_error
            - not_found
            - invalid_account
            - account_conflict
            - invalid_node_id
            - node_not_found
            - node_occupied
            - node_in_use
            - node_revision_conflict
            - node_assignment_conflict
            - node_name_conflict
            - platform_region_mismatch
            - invalid_node
            - invalid_exit_address
            - invalid_combination_id
            - combination_not_found
            - combination_occupied
            - combination_incompatible
            - combination_conflict
            - invalid_combination
            - resource_in_use
            - credential_required
            - credential_too_large
            - invalid_provider_id
            - provider_not_found
            - provider_revision_conflict
            - provider_conflict
            - invalid_provider
            - invalid_target_id
            - target_not_found
            - collection_not_found
            - collection_conflict
            - target_not_stopped
            - collection_in_use
            - invalid_collection
            - collection_storage_unavailable
            - collection_error
            - invalid_run_id
            - invalid_page_sequence
            - page_payload_not_found
            - invalid_limit
            - invalid_offset
            - invalid_price_range
            - invalid_appid
            - invalid_quotes
...
    Watermark:
      type: object
      additionalProperties: false
      required: [region, min_usable, usable, revision]
      properties:
        region: {type: string, enum: [domestic, foreign, hongkong]}
        min_usable: {type: integer, minimum: 0}
        usable: {type: integer, minimum: 0}
        revision: {type: integer, format: int64, minimum: 1}
    WatermarkSet:
      type: object
      additionalProperties: false
      required: [region, expected_revision, min_usable]
      properties:
        region: {type: string, enum: [domestic, foreign, hongkong]}
        expected_revision: {type: integer, format: int64, minimum: 1}
        min_usable: {type: integer, minimum: 0}
    Provider:
      type: object
      additionalProperties: false
      required: [id, name, enabled, priority, regions, has_credential, revision]
      properties:
        id: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        enabled: {type: boolean}
        priority: {type: integer, minimum: 1, maximum: 10000}
        regions:
          type: array
          items: {type: string, enum: [domestic, foreign, hongkong]}
        has_credential: {type: boolean}
        revision: {type: integer, format: int64, minimum: 1}
    ProviderCreate:
      type: object
      additionalProperties: false
      required: [name, priority, regions, credential]
      properties:
        name: {type: string}
        enabled: {type: boolean}
        priority: {type: integer, minimum: 1, maximum: 10000}
        regions:
          type: array
          items: {type: string, enum: [domestic, foreign, hongkong]}
        credential: {type: string, writeOnly: true, minLength: 1, maxLength: 65536}
    ProviderUpdate:
      type: object
      additionalProperties: false
      required: [expected_revision, name, enabled, priority, regions]
      properties:
        expected_revision: {type: integer, format: int64, minimum: 1}
        name: {type: string}
        enabled: {type: boolean}
        priority: {type: integer, minimum: 1, maximum: 10000}
        regions:
          type: array
          items: {type: string, enum: [domestic, foreign, hongkong]}
    ProviderCredentialReplace:
      type: object
      additionalProperties: false
      required: [expected_revision, credential]
      properties:
        expected_revision: {type: integer, format: int64, minimum: 1}
        credential: {type: string, writeOnly: true, minLength: 1, maxLength: 65536}
...
  responses:
    Error:
      description: Request rejected without reflecting request data or credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"